test: split e2e TAS tests into baseline and extended (#11002)
* test: split e2e TAS tests into baseline and extended

* test: make test-tas-e2e call baseline and extended targets

* docs: add label filter examples for TAS baseline and extended

* test: add helm targets for TAS baseline and extended

* test: update test-tas-e2e-helm to point to baseline helm target

* fix: add instanceType constant to helpers and remove from job_test.go

* test: remove old test/e2e/tas/ files

* test: move TAS tests to tas/baseline and tas/extended subfolders

* Update site/content/en/docs/contribution_guidelines/testing.md


* Update test/e2e/tas/baseline/suite_test.go


* Update site/content/en/docs/contribution_guidelines/testing.md


* Apply suggestion from @mbobrovskyi for tas files


* Apply suggestion from @mbobrovskyi


* Deleted old test declaration

* Remove unused nodes variable in suite_test.go

Remove unnecessary declaration of nodes variable.

* Update suite_test.go

* Reverted all the changes that were not required

Add additional setup steps for Kueue availability checks in the BeforeSuite function.

* Add NodeList initialization in suite_test.go

* Remove unused staticcheck exclusion rules

* Update test/e2e/tas/baseline/suite_test.go


* makinfg lint changes in suite_test.go file

* Revert "Remove unused staticcheck exclusion rules"

This reverts commit 3513abdc117607f8dd833fbe42fe5258bc8ac678.

* Reverted the changes in .golangci.yaml file and used lint-fix

* Revert "Reverted the changes in .golangci.yaml file and used lint-fix"

This reverts commit fc076c3a7a59b6f5ef317575d5dab236c8d17890.

* Made every issue okay as on 8 may by ekam-walia

* trying best to resolve conflicts on 8 may 21:56 IST

* 9 May 2026 first commit with change in suite_test.go

* Update test/e2e/tas/extended/suite_test.go


* Update test/e2e/tas/baseline/suite_test.go


* Final commit as on Saturday 9th May

---------
