+++ site/content/en/docs/contribution_guidelines/testing.md
@@ -57,7 +57,8 @@ For running a subset of tests, see [Running subset of tests](#running-subset-of-
 ```shell
 make kind-image-build
 make test-e2e
-make test-tas-e2e
+make test-tas-e2e-baseline
+make test-tas-e2e-extended
 make test-e2e-sequential-extended
 make test-e2e-sequential-baseline
 make test-e2e-certmanager
@@ -276,7 +277,8 @@ ginkgo.FIt("Should place pods based on the ranks-ordering", func() {
 and then run
 ```shell
 # build and pull image
-make test-tas-e2e
+make test-tas-e2e-baseline
+make test-tas-e2e-extended
 ```
 to test a particular TAS e2e test.
