+++ test/e2e/tas/extended/leaderworkerset_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package tas
+package extended
 
 import (
 	"fmt"
